// src/types/dto/in/PacketType.ts
enum PacketType {
  CHAT = "CHAT",
  STATE_CHANGE = "STATE_CHANGE",
  GAME_CREATED = "GAME_CREATED",
  GAME_REMOVED = "GAME_REMOVED",
  PLAYER_JOINED_GAME = "PLAYER_JOINED_GAME",
  PLAYER_LEFT_GAME = "PLAYER_LEFT_GAME",
  OBSERVER_JOINED_GAME = "OBSERVER_JOINED_GAME",
  OBSERVER_LEFT_GAME = "OBSERVER_LEFT_GAME",
  GAME_SETTINGS_UPDATED = "GAME_SETTINGS_UPDATED"
}

export default PacketType;